+++ src/index.js
@@ -7,3 +7,4 @@ import './js/btnScroll';
 import './js/paginationFromFetch';
 
 import './js/auth';
+import './js/modal';
