+++ src/partials/modal.html
@@ -1,27 +1,3 @@
 <div class="modal">
-<div class="modal__body container">
-        <img src="../src/images/image-demo.jpg" alt="постер">
-    <div class="modal__content">
-         <button type="button" class="modal__close">x</button>
-    <H2 class="modal__header">Заголовок</H2>
-   <ul class="modal-list list">
-     <li class="modal-list__item"><p class="modal-list__stat">Vote/Votes</p>
-      <p class="modal-list__value"> <span class="rating">6.3</span>/1200</p>
-    </li>
-       <li class="modal-list__item"><p class="modal-list__stat">Popularity</p>
-      <p class="modal-list__value">100.2</p>
-    </li>
-       <li class="modal-list__item"><p class="modal-list__stat">Original Title</p>
-      <p class="modal-list__value film-title">a wistful of lead</p>
-    </li>
-       <li class="modal-list__item"><p class="modal-list__stat">Genre</p>
-      <p class="modal-list__value">Western</p>
-    </li>
-    </ul>
-<p>About</p>
-    <p class="description">Lorem ipsum dolor sit amet consectetur adipisicing elit. Reprehenderit eius excepturi ipsum obcaecati odit doloribus sapiconsequuntur optio autem assumenda cum?</p>
-      <button type="button">add to Watched</button>
-    <button type="button">add to queue</button>
-  </div>
-</div>
+    <button type="button" class="modal__close">x</button>
 </div>
