+++ src/sass/layout/_modal-window.scss
@@ -1,67 +1,54 @@
 .modal {
-  position: fixed;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0,0,0,0.25);
-  opacity: 0;
-  visibility: hidden;
-};
-
-.modal__body {
-  z-index: 1;
   display: flex;
   align-items: center;
   justify-content: center;
-  background-color: #fefefe;
-  margin: 15% auto;
-  padding: 20px;
-  width: 80%;
-};
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  opacity: 0;
+  pointer-events: none;
+  transition: opacity cubic-bezier(0.4, 0, 0.2, 1) 250ms;
+  z-index: 2;
+}
+
+.modal.is-open {
+  opacity: 1;
+  pointer-events: initial;
+}
+
+body.modal-open {
+  overflow: hidden;
+}
+
+.modal-backdrop {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.25);
+}
+// Контейнер модалки
+.modal__container {
+  overflow-x: auto;
+  position: relative;
+
+  max-height: 100vh;
+  width: 280px;
+  margin: auto;
+  padding: 48px 20px 41px 20px;
+  border-radius: 6px;
+  background-color: #ffffff;
+  color: #000000;
+  z-index: 2;
+  font-family: Roboto, sans-serif;
+}
 
 .modal__close {
-  float: right;
-};
-
-.modal__header {
-  color: #000000;
-  font-size: 30px;
-  font-weight: 500;
-  text-transform: uppercase;
-};
-
-.modal-list__stat {
-font-weight: 500;
-font-size: 12px;
-line-height: 1.33;
-color: #8C8C8C;
-};
-
-.modal-list__value {
-  font-weight: 500;
-font-size: 12px;
-line-height: 14px;
-  color: #000000;
-};
-
-.rating {
-color: white;
-background: #FF6B01;
-border-radius: 5px;
-};
-
-.film-title {
-  text-transform: uppercase;
-};
-
-.description {
-font-weight: 500;
-font-size: 12px;
-line-height: 1.67;
-color: #000000;
-};
-
-.modal__content {
-  padding-left: 20px;
-}
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  z-index: 99;
+ }
